==> 01_nlp_stopwords_punct_entity.ipynb <==
# NLP: Initial Text Pre-Processing

- Remove stop words
- Remove numbers
- Remove punctuation
- Underscore entities with more than one word
- Tidy up Reuters news article text

## Imports

import pandas as pd
import numpy as np
import os
import re

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.decomposition import TruncatedSVD, NMF
from sklearn.preprocessing import Normalizer, normalize
from sklearn import metrics
from sklearn.metrics import silhouette_score
from sklearn.cluster import KMeans, MiniBatchKMeans, AgglomerativeClustering, DBSCAN, SpectralClustering
from sklearn.utils.extmath import randomized_svd
from sklearn.pipeline import make_pipeline
from sklearn.neighbors import NearestNeighbors
from sklearn.decomposition import LatentDirichletAllocation
from sklearn.mixture import GaussianMixture
from sklearn.model_selection import train_test_split

import nltk
from nltk.stem.porter import PorterStemmer
from nltk.chunk import ne_chunk
from nltk.tag import pos_tag
from nltk.tokenize import word_tokenize
from nltk.stem.lancaster import LancasterStemmer
from nltk.corpus import stopwords

from nltk.tag import StanfordNERTagger
from string import punctuation
from gensim.models.phrases import Phrases, Phraser
from gensim import corpora, models, similarities, matutils, models
import spacy

import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline

## Config

java_path = "C:/Program Files/Java/jdk1.8.0_144/bin/java.exe"
os.environ['JAVAHOME'] = java_path

stanford_ner_path = 'C:\ds\stanford-ner\stanford-ner.jar'
stanford_classifier = 'C:\ds\stanford-ner\classifiers\english.muc.7class.distsim.crf.ser.gz'

nlp = spacy.load(name='en')

stops = stopwords.words('english')

extra_stops = ['pct', 'percent', 'cent', 'high', 'low', 'top', 'news', 'topnew', 'topnews', 'GMT', 'BST', 'AM', 'PM',
              'Reuters', 'reuters', 'reuters.com', 'plc', 'PLC', 'visit', 'click', 'thomson', 'Thomson', 'thomsonreuters',
              'suggest', 'feedback', 'alert', 'email', 'best', 'cms', 'CMS', 'pageid', 'livemarket', 'client', 'link',
              'net', 'change', 'chg', 'open', 'site', 'eikon', 'EIKON', 'yld', 'yr', 'say', 'year', 'close', 'performance',
              'perform', 'performs', 'id', 'ID', 'pa', 'report', 'reports', 'reporting', 'share', 'break', 'recent', 'past',
              'point', 'said', 'index', 'data', 'new', 'points', 'market', 'markets', 'bn', 'Bn', 'Mn', 'mn', 'avg', 'average',
              'fell', 'fall', 'rose', 'rise', 'time', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 
               'Sunday', 'people', 'higher', 'lower', 'likely', 'Reporting']

stops = stops + extra_stops
stops = set(stops)

## Functions

def underscore_entities(document):
    spacy = nlp(document)
    
    uds_entities = []
    ent_chars = []
    for i in spacy.ents:
        ent_chars.append((i.start_char, i.end_char))
    
    new_string = []
    for index, item in enumerate(list(str(spacy))):
        for i, j in enumerate(ent_chars):
            if j[0] < index < j[1] and item == ' ':
                item = '_'
        new_string.append(item)

    ns = ''.join(new_string)
    uds_entities.append(ns)
        
    return uds_entities

def clean_reuters(article):
    text_to_find = '(Reuters) - '
    try:
        cleaned = article[article.index(text_to_find) + len(text_to_find):]
        return cleaned
    except:
        return article

def clean_text(raw_text, stop=True):
    raw_text = raw_text.replace('U.S', 'US')
    raw_text = clean_reuters(raw_text)
    letters_only = re.sub('[^a-zA-Z]', ' ', raw_text)
    letters_only = ' '.join(letters_only.split())
#     underscored = underscore_entities(letters_only)
#     words = underscored[0].split()
    words = letters_only.split()
    
    if stop == True:
        meaningful_words = [w for w in words if not w in stops]
        return( " ".join(meaningful_words))
    else:
        return( " ".join(words))

## Load Data

df = pd.read_csv('news_articles.csv')

df.info()

## Clean Article Text

df['cleaned_article_text'] = df['article_text'].apply(clean_text)

df.sample(10)

## Save Data

df.to_csv('news_articles_cleaned.csv', index=None, encoding='utf-8')

df.iloc[35750]['article_text']

df.iloc[35750]['cleaned_article_text']
