==> 02_trim_cleaned_article_data.ipynb <==
# NLP: Trim Article Dataset

## Imports

import pandas as pd

## Load Data

df = pd.read_csv('news_articles_cleaned.csv')

df.info()

## Remove articles < 150 characters

df.sample(10)

mask = (df['cleaned_article_text'].str.len() > 150)
df = df[mask]

df.info()

## Save Data

df.to_csv('news_articles_cleaned_trimmed.csv', index=None, encoding='utf-8')
