==> 03_nlp_lda_extra_stopwords.ipynb <==
# NLP: Use LDA to identify additional stop words

## Imports

import pandas as pd
from nltk.corpus import stopwords
from gensim import corpora, models, similarities, matutils
from sklearn.feature_extraction.text import CountVectorizer

## Functions

def remove_stop_words(raw_text, stop_list):
    words = raw_text.split()
    cleaned = [w for w in words if not w in stop_list]
    return( " ".join(cleaned))

## Load Data

df = pd.read_csv('news_articles_cleaned_trimmed.csv')

df.sample(5)

## LDA

Take a sample from the entire dataset.

sample_df = df.sample(10000)

count_vectorizer = CountVectorizer(stop_words='english')
count_vectorizer.fit(sample_df['cleaned_article_text'])

counts = count_vectorizer.transform(sample_df['cleaned_article_text']).transpose()

counts.shape

corpus = matutils.Sparse2Corpus(counts)

id2word = dict((v, k) for k, v in count_vectorizer.vocabulary_.items())

len(id2word)

lda = models.LdaMulticore(corpus=corpus, num_topics=10, id2word=id2word, passes=1)

lda.print_topics()

## Remove extra stop words

extra_stopwords = ['shares', 'economy', 'economic', 'prices', 'investors']

stop_list = stopwords.words('english')
len(stop_list)

stop_list = stop_list + extra_stopwords
len(stop_list)

sample_df['article_text_extra_stops'] = sample_df['cleaned_article_text'].apply(remove_stop_words, stop_list=stop_list)

sample_df.sample(5)

len(sample_df.iloc[2028]['cleaned_article_text'])

len(sample_df.iloc[2028]['article_text_extra_stops'])

## Repeat process

### LDA

count_vectorizer = CountVectorizer(stop_words='english')
count_vectorizer.fit(sample_df['article_text_extra_stops'])
counts = count_vectorizer.transform(sample_df['article_text_extra_stops']).transpose()
counts.shape

corpus = matutils.Sparse2Corpus(counts)
id2word = dict((v, k) for k, v in count_vectorizer.vocabulary_.items())
len(id2word)

lda = models.LdaMulticore(corpus=corpus, num_topics=10, id2word=id2word, passes=1)
lda.print_topics()

### Stop words

extra_stopwords = extra_stopwords + ['government', 'growth', 'banks']
len(extra_stopwords)

stop_list = stopwords.words('english')
stop_list = stop_list + extra_stopwords
len(stop_list)

sample_df['article_text_extra_stops_v2'] = sample_df['cleaned_article_text'].apply(remove_stop_words, stop_list=stop_list)

sample_df.sample(5)

print(len(sample_df.iloc[2028]['cleaned_article_text']))
print(len(sample_df.iloc[2028]['article_text_extra_stops']))
print(len(sample_df.iloc[2028]['article_text_extra_stops_v2']))

## And again..

### LDA

count_vectorizer = CountVectorizer(stop_words='english')
count_vectorizer.fit(sample_df['article_text_extra_stops_v2'])
counts = count_vectorizer.transform(sample_df['article_text_extra_stops_v2']).transpose()
counts.shape

corpus = matutils.Sparse2Corpus(counts)
id2word = dict((v, k) for k, v in count_vectorizer.vocabulary_.items())
len(id2word)

lda = models.LdaMulticore(corpus=corpus, num_topics=10, id2word=id2word, passes=1)
lda.print_topics()

### Stop words

extra_stopwords = extra_stopwords + ['expected', 'rate', 'rates', 'stocks', 'financial']
len(extra_stopwords)

stop_list = stopwords.words('english')
stop_list = stop_list + extra_stopwords
len(stop_list)

sample_df['article_text_extra_stops_v3'] = sample_df['cleaned_article_text'].apply(remove_stop_words, stop_list=stop_list)

sample_df.sample(5)

print(len(sample_df.iloc[2028]['cleaned_article_text']))
print(len(sample_df.iloc[2028]['article_text_extra_stops']))
print(len(sample_df.iloc[2028]['article_text_extra_stops_v2']))
print(len(sample_df.iloc[2028]['article_text_extra_stops_v3']))

## Again..

### LDA

count_vectorizer = CountVectorizer(stop_words='english')
count_vectorizer.fit(sample_df['article_text_extra_stops_v3'])
counts = count_vectorizer.transform(sample_df['article_text_extra_stops_v3']).transpose()
counts.shape

corpus = matutils.Sparse2Corpus(counts)
id2word = dict((v, k) for k, v in count_vectorizer.vocabulary_.items())
len(id2word)

lda = models.LdaMulticore(corpus=corpus, num_topics=10, id2word=id2word, passes=1)
lda.print_topics()

### Stop words

extra_stopwords = extra_stopwords + ['uk', 'european', 'u_s', 'tax', 'britain', 'inflation', 'country', 'bank', 
                                    'business', 'spending', 'dollar', 'oil', 'company', 'stock', 'yen', 'fed', 'earnings',
                                    'sector', 'global', 'policy', 'world', 'debt', 'china', 'biggest', 'trade', 'bank',
                                    'work', 'nikkei', 'profit', 'business', 'European', 'UK', 'U_S', 'Bank', 'Nikkei', 'Fed', 
                                     'Britain', 'Dollar', 'Yen', 'Global', 'World', 'China']
len(extra_stopwords)

stop_list = stopwords.words('english')
stop_list = stop_list + extra_stopwords
len(stop_list)

stop_list = set(stop_list)
len(stop_list)

sample_df['article_text_extra_stops_v4'] = sample_df['cleaned_article_text'].apply(remove_stop_words, stop_list=stop_list)

print(len(sample_df.iloc[2000]['cleaned_article_text']))
print(len(sample_df.iloc[2000]['article_text_extra_stops']))
print(len(sample_df.iloc[2000]['article_text_extra_stops_v2']))
print(len(sample_df.iloc[2000]['article_text_extra_stops_v3']))
print(len(sample_df.iloc[2000]['article_text_extra_stops_v4']))

## Again

### LDA

count_vectorizer = CountVectorizer(stop_words='english')
count_vectorizer.fit(sample_df['article_text_extra_stops_v4'])
counts = count_vectorizer.transform(sample_df['article_text_extra_stops_v4']).transpose()
counts.shape

corpus = matutils.Sparse2Corpus(counts)
id2word = dict((v, k) for k, v in count_vectorizer.vocabulary_.items())
len(id2word)

lda = models.LdaMulticore(corpus=corpus, num_topics=10, id2word=id2word, passes=1)
lda.print_topics()

### Stop words

extra_stopwords = extra_stopwords + ['Euro', 'euro', 'EU', 'eu', 'Europe', 'europe', 'long', 'Japan', 'japan', 'says', 
                                    'hit', 'reporting', 'Reporting']
len(extra_stopwords)

stop_list = stopwords.words('english')
stop_list = stop_list + extra_stopwords
len(stop_list)

sample_df['article_text_extra_stops_v5'] = sample_df['cleaned_article_text'].apply(remove_stop_words, stop_list=stop_list)

print(len(sample_df.iloc[2000]['cleaned_article_text']))
print(len(sample_df.iloc[2000]['article_text_extra_stops']))
print(len(sample_df.iloc[2000]['article_text_extra_stops_v2']))
print(len(sample_df.iloc[2000]['article_text_extra_stops_v3']))
print(len(sample_df.iloc[2000]['article_text_extra_stops_v4']))
print(len(sample_df.iloc[2000]['article_text_extra_stops_v5']))

## Again

### LDA

count_vectorizer = CountVectorizer(stop_words='english')
count_vectorizer.fit(sample_df['article_text_extra_stops_v5'])
counts = count_vectorizer.transform(sample_df['article_text_extra_stops_v5']).transpose()
counts.shape

corpus = matutils.Sparse2Corpus(counts)
id2word = dict((v, k) for k, v in count_vectorizer.vocabulary_.items())
len(id2word)

lda = models.LdaMulticore(corpus=corpus, num_topics=10, id2word=id2word, passes=1)
lda.print_topics()

### Stop words

extra_stopwords = extra_stopwords + ['strong', 'financial', 'Strong', 'Financial']

['shares',
 'economy',
 'economic',
 'prices',
 'investors',
 'government',
 'growth',
 'banks',
 'expected',
 'rate',
 'rates',
 'stocks',
 'financial',
 'uk',
 'european',
 'u_s',
 'tax',
 'britain',
 'inflation',
 'country',
 'bank',
 'business',
 'spending',
 'dollar',
 'oil',
 'company',
 'stock',
 'yen',
 'fed',
 'earnings',
 'sector',
 'global',
 'policy',
 'world',
 'debt',
 'china',
 'biggest',
 'trade',
 'bank',
 'work',
 'nikkei',
 'profit',
 'business',
 'uk',
 'european',
 'u_s',
 'tax',
 'britain',
 'inflation',
 'country',
 'bank',
 'business',
 'spending',
 'dollar',
 'oil',
 'company',
 'stock',
 'yen',
 'fed',
 'earnings',
 'sector',
 'global',
 'policy',
 'world',
 'debt',
 'china',
 'biggest',
 'trade',
 'bank',
 'work',
 'nikkei',
 'profit',
 'business',
 'uk',
 'european',
 'u_s',
 'tax',
 'britain',
 'inflation',
 'country',
 'bank',
 'business',
 'spending',
 'dollar',
 'oil',
 'company',
 'stock',
 'yen',
 'fed',
 'earnings',
 'sector',
 'global',
 'policy',
 'world',
 'debt',
 'china',
 'biggest',
 'trade',
 'bank',
 'work',
 'nikkei',
 'profit',
 'business',
 'European',
 'UK',
 'U_S',
 'Bank',
 'Nikkei',
 'Fed',
 'Britain',
 'Dollar',
 'Yen',
 'Global',
 'World',
 'China',
 'Euro',
 'euro',
 'EU',
 'eu',
 'Europe',
 'europe',
 'long',
 'Japan',
 'japan',
 'says',
 'hit',
 'reporting',
 'Reporting']

extra_stopwords = list(set(extra_stopwords))

## Apply extra stop word removal to full data set

stop_list = stopwords.words('english')
stop_list = stop_list + extra_stopwords
len(stop_list)

df['article_text_extra_stops'] = df['cleaned_article_text'].apply(remove_stop_words, stop_list=stop_list)

df.sample(5)

## Save data

df.to_csv('news_articles_extra_stops.csv', index=None, encoding='utf-8')
