==> 04_nlp_stem_lemmatize_text.ipynb <==
# NLP: Stemming & Lemmatizing

## Imports

import pandas as pd

import nltk
from nltk.stem.porter import PorterStemmer
from nltk.stem.lancaster import LancasterStemmer
from nltk.stem import WordNetLemmatizer, SnowballStemmer

## Functions

def stem_lem_text(s, type='Lancaster'):
    words = s.split()
    
    if type == 'Porter':
        choice = PorterStemmer()
        reformed = [choice.stem(word) for word in words]
    elif type == 'Snowball':
        choice = SnowballStemmer('english')
        reformed = [choice.stem(word) for word in words]
    elif type == 'Lemmatize':
        choice = WordNetLemmatizer()
        reformed = [choice.lemmatize(word) for word in words]
    else:
        choice = LancasterStemmer()
        reformed = [choice.stem(word) for word in words]
        
    reformed = " ".join(reformed)
    return reformed

## Load Data

df = pd.read_csv('news_articles_cleaned_trimmed.csv')

df.info()

df.sample(10)

## Stem / Lemmatize Text

df['text_lancaster'] = df['cleaned_article_text'].apply(stem_lem_text, type='Lancaster')
df['text_porter'] = df['cleaned_article_text'].apply(stem_lem_text, type='Porter')
df['text_snowball'] = df['cleaned_article_text'].apply(stem_lem_text, type='Snowball')
df['text_lemmatize'] = df['cleaned_article_text'].apply(stem_lem_text, type='Lemmatize')

## Save Data

df.to_csv('news_articles_stemmed_lemmatized_no_extra_stops.csv', index=None, encoding='utf-8')
